requirements.txt
altgraph==0.17.4
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
customtkinter==5.2.2
darkdetect==0.8.0
Django==5.1.6
Flask==3.1.0
h11==0.14.0
idna==3.6
image==1.5.33
itsdangerous==2.2.0
Jinja2==3.1.4
jsonify==0.5
lxml==5.3.0
MarkupSafe==3.0.2
outcome==1.3.0.post0
packaging==23.2
pefile==2023.2.7
pillow==11.1.0
pycparser==2.21
pyinstaller==6.4.0
pyinstaller-hooks-contrib==2024.1
PySocks==1.7.1
pywin32-ctypes==0.2.2
requests==2.31.0
selenium==4.17.2
setuptools==69.1.0
six==1.17.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2025.1
urllib3==2.2.0
Werkzeug==3.1.3
wsproto==1.2.0
